requirements.txt
# ─── Core ML ─────────────────────────────────────────────────────────────────
torch==2.2.2
torchvision==0.17.2
timm==0.9.16

# ─── Data & Numerics ──────────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
Pillow==10.3.0

# ─── Visualization ────────────────────────────────────────────────────────────
matplotlib==3.8.4
seaborn==0.13.2

# ─── Web Scraping ─────────────────────────────────────────────────────────────
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.1

# ─── Retrieval ────────────────────────────────────────────────────────────────
thefuzz==0.22.1
python-Levenshtein==0.25.1

# ─── UI ───────────────────────────────────────────────────────────────────────
streamlit==1.35.0

# ─── Config & Utilities ───────────────────────────────────────────────────────
PyYAML==6.0.2
tqdm==4.66.4

# ─── Image Download ───────────────────────────────────────────────────────────
icrawler==0.6.6

# ─── Testing ──────────────────────────────────────────────────────────────────
pytest==8.2.2

# ─── OCR (requires: brew install tesseract) ───────────────────────────────────
pytesseract==0.3.10
